requirements.txt
asn1crypto==0.24.0
aspy.yaml==1.1.1
atomicwrites==1.2.1
attrs==18.2.0
cached-property==1.5.1
certifi==2018.10.15
cffi==1.11.5
cfgv==1.1.0
chardet==3.0.4
Click==7.0
cnx-litezip==1.5.0
cnxml==2.2.0
cryptography==2.4.1
filelock==3.0.10
github3.py==1.2.0
identify==1.1.7
idna==2.7
importlib-metadata==0.6
jwcrypto==0.6.0
lxml==4.2.5
more-itertools==4.3.0
nebuchadnezzar==6.1.0
nodeenv==1.3.3
pluggy==0.8.0
pre-commit==1.12.0
py==1.7.0
PyPOM==2.2.0
pycparser==2.19
pytest==3.10.1
pytest-base-url==1.4.1
pytest-html==1.19.0
pytest-metadata==1.7.0
pytest-selenium==1.14.0
pytest-testrail==2.3.1
pytest-variables==1.7.1
python-dateutil==2.7.5
python-dotenv==0.9.1
python-magic==0.4.15
PyYAML==3.13
requests==2.20.1
selenium==3.14.1
simplejson==3.16.0
six==1.11.0
toml==0.10.0
tox==3.5.3
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.1.0
zope.component==4.5
zope.deferredimport==4.3
zope.deprecation==4.3.0
zope.event==4.4
zope.hookable==4.2.0
zope.interface==4.6.0
zope.proxy==4.3.1
